backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
pytest==7.4.2
requests==2.31.0
PyJWT==2.8.0
SQLAlchemy==2.0.21
jsonschema==4.19.0
PyYAML==6.0.1
Faker==19.13.0

